Se terminan funcionalidades

+++ src/main/java/pch/huellaschile/msusuariosmascotas/application/services/mascota/DeleteMascotaService.java
@@ -1,7 +1,9 @@
 package pch.huellaschile.msusuariosmascotas.application.services.mascota;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 import pch.huellaschile.msusuariosmascotas.application.ports.input.mascota.DeleteMascotaUseCase;
 import pch.huellaschile.msusuariosmascotas.application.ports.output.mascota.MascotaGateway;
 
@@ -13,6 +15,12 @@ public class DeleteMascotaService implements DeleteMascotaUseCase {
 
     @Override
     public Boolean execute(Integer idMascota){
+        boolean exist = gateway.existsByIdMascotaAndTratamientoTrue(idMascota);
+
+        if(exist) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Tiene tratamiento activo");
+        }
+
         return gateway.getMascotaById(idMascota).map(certification -> {
             gateway.deleteMascotaById(idMascota);
             return true;
